fwmod: update /usr/bin/system_status correctly even when FREETZ_SUBVERSION_STRING is not selected. rudishell: show it only with security level 0, don't allow to execute any code by clicking on a wrong url (closes #294) webcfg: add possibility to change username (under 'settings' - 'webcfg', closes #255)
git-svn-id: http://svn.freetz.org/trunk@3018 f5190166-0702-4917-9039-51ec32eddaf5

+++ root/usr/mww/cgi-bin/rudi_upload.cgi
@@ -1,6 +1,15 @@
 #!/usr/bin/haserl -u 10000 -U /var/tmp
 Content-Type: text/html; charset=ISO-8859-1
-$(lang de:"test de" en:"test en")
+
+<?
+pid1=`echo ${QUERY_STRING} | sed -n 's/.*pid=\(.*\)/\1/p' | sed -e 's/&.*//g'`
+pid2=`cat /var/run/rudi_shell.pid`
+if [ "$pid1" != "$pid2" ] || [ ! -f /var/run/rudi_shell.pid ]; then
+	echo -n '<html><body>$(lang de:"Sicherheitsabfrage fehlgeschlagen! Falsche PID!" en:"Security check failed! Wrong PID!")</body></html>'
+	exit 1
+fi
+?>
+
 <html><body>
 	<pre id="cmd_output"><? if mv "$FORM_source" "$FORM_target"
 		then
